Think have memory allocation working. Also expand file name in Windows

+++ def_glob.h
@@ -31,14 +31,17 @@ EXTERN char *arg_v[ARGNN]; // Modified DJD
 
 */
 
-EXTERN int alloc_words;
+
+// EXTERN int alloc_words;
+EXTERN unsigned long alloc_words;
 
 /*! \var mem_size
   \brief number of words from memory = alloc_words * sizeof word 
 
 */
 
-EXTERN int mem_size;
+// EXTERN int mem_size;
+EXTERN unsigned long mem_size;
 
 /*! \var mem_base
   \brief mem_size memory allocated in init_memory by malloc
